Fix queries, these have been validated as functional by client impl (#339)
* Fix queries, these have been validated as functional by client impl

* change of not found status code for who is queries
